+++ packages/components/CHANGELOG.md
@@ -14,9 +14,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Changed
+
+- All components (expect some form elements) have improved focus styling based on `:focus-visible`
+  ([#3011](https://github.com/porsche-design-system/porsche-design-system/pull/3011))
+- Several components are using CSS property `inset|inset-inline|inset-block` instead of `top|bottom|left|right` for
+  better RTL (right-to-left) support ([#3011](https://github.com/porsche-design-system/porsche-design-system/pull/3011))
+- `Switch`: Improve RTL (right-to-left) mode
+  ([#2992](https://github.com/porsche-design-system/porsche-design-system/pull/2992))
 - `Button`, `Button Pure`, `Switch`, `Checkbox Wrapper`, `Radio Button Wrapper`, `Pin Code`: optimized announcement of
   loading state for assistive technologies
   ([#3009](https://github.com/porsche-design-system/porsche-design-system/pull/3009))
+
+#### Fixed
+
+- All components are supporting focus style in High Contrast Mode correctly
+  ([#3011](https://github.com/porsche-design-system/porsche-design-system/pull/3011))
 
 ### [3.11.0] - 2024-01-30